Bump eslint version 0.15.0 -> 0.16.0
This new version of eslint contains many improvements, including better
support for ES6 features. The full changelog can be found at:

  https://github.com/eslint/eslint/blob/master/CHANGELOG.md

With this change, I ran into an issue that was causing the tests to
pass. Lots of errors like:

> 10:12  error  "fs" used outside of binding context  block-scoped-var

It appears that this is an open issue with 0.16.0 that can be worked
around by setting `globalReturn` to `false` in the `.eslintrc` file.
More information:

  https://github.com/eslint/eslint/issues/1969