jsx-first-prop-new-line: add multiline-multiprop option
This new option is similar to `multiline`, but it will only enforce a
new line if the tag spans multiple lines *and* are more than 1 prop.

Also, clear up the language in the docs for the `multiline` option. We
are checking that the tag spans multiple lines. This does not imply
there are multiple props (1 prop can span multiple lines).