add props to lifecycle sort group
When using Flow, you’ll often include types for your props like so:

```
class MyComponent extends Component {
  props: {
    a: number,
    b: number,
  }

  state: {
    c: number,
    d: number,
  } = {
    c: 1,
    d: 2,
  }

  ...
}
```

In such a case you will also want the `props` type definition on top of the component. This PR adds the `props` to the `sort-comp` `lifecycle` rule so that by default the `props` property will be sorted to the top. Adding a `props` property is generally not desirable in any other case but this one.