Add forbidPatterns to forbid-component-props rule
Adds a forbidPatterns option that takes an array
of safe pattern strings so it's possible to forbid the use
of things like "data-*" props.
Also adds the ignoreDomNodes option (defaults to true), because
sometimes we do want to check forbidden props on DOM nodes
i.e. for "data-*" so we can avoid tying implementation
to thirdparty libraries that work on the DOM.