Add no-array-index-key rule
This rule will warn when using the array index as the `key`. As a first
pass, I decided to implement this for `Array.prototype.map`, since that
is likely the most common case, but it theoretically could be expanded
to find other cases so I kept the naming more generic.

Like many rules, this one is imperfect and is prone to some false
positives and negatives. For instance, if someone defines a `.map()`
function on an object that doesn't have the same signature as
`Array.prototype.map`, this will likely end up warning in those cases.
However, I think the value of this rule outweighs its hypothetical
drawbacks.