Fix crash in no-access-state-in-setstate
With an IIFE in the return statement of render that uses `this.state`,
there is no parent with a name. In this rule, the method names are later
checked to see if they are being called elsewhere in the file. Since
there is no intention of this being called anywhere else, the IIFE will
just be skipped being added to the list of methods.