Allow jsx-pascal-case to ignore using regexps
We want to ignore a consistent theme of component naming which matches a
regexp, without having to individually label each of the components, or
to individually eslint-disable each of the instances.

This should be backwards compatible with existing rule sets, and matches
how the `camelcase` rule works in eslint (which also supports regexps).