fix: add additional check to ensure callee object name is `PropTypes`
An issue was raised where using an alternative library with a similar syntax (in this case Yup) resulted in the rule `forbid-prop-types` stating that `Yup.array()` was an invalid proptype. Adding another check to ensure pragma is used and the check is off of `PropTypes`, removes the false erorrs thrown up when using those other libraries.

Signed-off-by: Josh Kelly <josh@codecog.dev>