[Fix] `jsx-no-literals`: properly handle HTML entities when allowed
When an HTML entity such as &mdash; is checked with ESLint, the `value`
property becomes the actual unicode character instead of the raw
literal.

This means if `&mdash;` is in the allowed strings array, and `&mdash;`
is in the code being checked, the actual value used in the JSXText check
will be `—`, which will not match against `&mdash;`, so it causes a
false positive error.

When checking against the set of allowed strings, this will now check
both the `value` and the `raw` properties of the node. For `&mdash;`, the
`raw` property will still be `&mdash;`.

If _either_ of these match the list of allowed values, then no error
will be returned.

Fixes #3130

Co-authored-by: Joe Attardi <attardi@synopsys.com>
Co-authored-by: Ross <52366381+ROSSROSALES@users.noreply.github.com>